requirements.txt
azure-core
azure-eventhub
azure-storage-blob
azure-functions
azure-identity
